Split ledger tests under line limits
