+++ crates/axon-api/src/source_tests.rs
@@ -491,32 +491,5 @@ fn capability_document_uses_closed_provider_enums() {
     assert_eq!(value["adapters"][0]["owner_crate"], "axon-adapters");
 }
 
-#[test]
-fn document_cache_invalidation_rejects_unknown_variant_fields() {
-    for payload in [
-        serde_json::json!({
-            "kind": "source",
-            "source_id": "src",
-            "extra": true
-        }),
-        serde_json::json!({
-            "kind": "generation",
-            "generation": "gen",
-            "extra": true
-        }),
-        serde_json::json!({
-            "kind": "key",
-            "key": {
-                "source_id": "src",
-                "source_item_key": "item"
-            },
-            "extra": true
-        }),
-        serde_json::json!({
-            "kind": "all",
-            "extra": true
-        }),
-    ] {
-        assert!(serde_json::from_value::<DocumentCacheInvalidation>(payload).is_err());
-    }
-}
+#[path = "source_tests/document_cache_tests.rs"]
+mod document_cache_tests;
