+++ crates/axon-ledger/src/store_tests.rs
@@ -1,5 +1,4 @@
 use axon_api::source::*;
-use uuid::Uuid;
 
 use crate::store::{FakeLedgerStore, LedgerStore};
 
@@ -182,532 +181,9 @@ fn lease_request_ttl(key: &str, owner: &str, ttl_seconds: u64) -> LeaseRequest {
     }
 }
 
-#[tokio::test]
-async fn fake_ledger_diffs_manifests_and_tracks_committed_generation() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-
-    let first = ledger.diff_manifest(manifest("a")).await.unwrap();
-    assert_eq!(first.counts.added, 1);
-    let first_manifest = manifest("a");
-    ledger.put_manifest(first_manifest.clone()).await.unwrap();
-
-    let generation = completed_generation_for_manifest(&first_manifest);
-    let generation = complete_and_publish(&ledger, generation.clone()).await;
-    assert_eq!(generation.publish_state, PublishState::Committed);
-    assert!(generation.published_at.is_some());
-
-    let refreshed = ledger.diff_manifest(manifest("b")).await.unwrap();
-    assert_eq!(refreshed.counts.modified, 1);
-    assert_eq!(
-        ledger.committed_generation(&SourceId::new("src_a")).await,
-        Some(generation.generation)
-    );
-}
-
-#[tokio::test]
-async fn fake_ledger_diffs_only_against_committed_generation() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-
-    ledger.put_manifest(manifest("uncommitted")).await.unwrap();
-
-    let diff = ledger.diff_manifest(manifest("next")).await.unwrap();
-    assert_eq!(diff.previous_generation, None);
-    assert_eq!(diff.counts.added, 1);
-    assert_eq!(diff.counts.modified, 0);
-    assert_eq!(diff.counts.unchanged, 0);
-}
-
-#[tokio::test]
-async fn fake_rejects_invalid_manifest_item_ownership_and_duplicates() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-
-    let mut wrong_source = manifest("wrong-source");
-    wrong_source.items[0].source_id = SourceId::new("other");
-    let error = ledger.put_manifest(wrong_source).await.unwrap_err();
-    assert_eq!(
-        error.code.to_string(),
-        "source.ledger.manifest_item_source_mismatch"
-    );
-
-    let mut duplicate = manifest_with_items(
-        "gen_duplicate",
-        vec![
-            manifest_item("src/lib.rs", "a"),
-            manifest_item("src/lib.rs", "b"),
-        ],
-    );
-    duplicate.items[1].canonical_uri = "file:///repo/src/lib-copy.rs".to_string();
-    let error = ledger.put_manifest(duplicate).await.unwrap_err();
-    assert_eq!(
-        error.code.to_string(),
-        "source.ledger.manifest_duplicate_item"
-    );
-}
-
-#[tokio::test]
-async fn fake_ledger_scopes_generation_ids_per_source() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-    let mut src_b = source();
-    src_b.source_id = SourceId::new("src_b");
-    ledger.upsert_source(src_b).await.unwrap();
-
-    let src_a_first = ledger
-        .create_generation(SourceId::new("src_a"))
-        .await
-        .unwrap();
-    let src_b_first = ledger
-        .create_generation(SourceId::new("src_b"))
-        .await
-        .unwrap();
-    assert_eq!(src_a_first.generation, SourceGenerationId::new("gen_1"));
-    assert_eq!(src_b_first.generation, SourceGenerationId::new("gen_1"));
-
-    ledger
-        .put_manifest(manifest_for_generation(&src_a_first, "src-a-first"))
-        .await
-        .unwrap();
-    complete_and_publish(&ledger, completed_generation(src_a_first.clone())).await;
-    let src_a_second = ledger
-        .create_generation(SourceId::new("src_a"))
-        .await
-        .unwrap();
-    assert_eq!(src_a_second.generation, SourceGenerationId::new("gen_2"));
-    assert_eq!(
-        src_a_second.previous_generation,
-        Some(src_a_first.generation)
-    );
-}
-
-#[tokio::test]
-async fn fake_ledger_diffs_version_and_mtime_changes() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-    let previous = manifest_with_freshness("a", Some("v1"), ts());
-    ledger.put_manifest(previous.clone()).await.unwrap();
-    complete_and_publish(&ledger, completed_generation_for_manifest(&previous)).await;
-
-    let version_changed = ledger
-        .diff_manifest(manifest_with_freshness("a", Some("v2"), ts()))
-        .await
-        .unwrap();
-    assert_eq!(version_changed.counts.modified, 1);
-    assert_eq!(version_changed.counts.unchanged, 0);
-
-    let mtime_changed = ledger
-        .diff_manifest(manifest_with_freshness(
-            "a",
-            Some("v1"),
-            Timestamp("2026-07-02T00:00:00Z".to_string()),
-        ))
-        .await
-        .unwrap();
-    assert_eq!(mtime_changed.counts.modified, 1);
-    assert_eq!(mtime_changed.counts.unchanged, 0);
-}
-
-#[tokio::test]
-async fn fake_ledger_rejects_non_publishable_generation_statuses() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-    let running = ledger
-        .create_generation(SourceId::new("src_a"))
-        .await
-        .unwrap();
-
-    let error = ledger
-        .publish_generation(publish_request(&running))
-        .await
-        .unwrap_err();
-    assert_eq!(
-        error.code.to_string(),
-        "source.ledger.generation_not_publishable"
-    );
-    assert_eq!(
-        ledger.committed_generation(&SourceId::new("src_a")).await,
-        None
-    );
-
-    ledger
-        .put_manifest(manifest_for_generation(&running, "running"))
-        .await
-        .unwrap();
-    complete_and_publish(&ledger, completed_generation(running.clone())).await;
-    assert_eq!(
-        ledger.committed_generation(&SourceId::new("src_a")).await,
-        Some(running.generation)
-    );
-}
-
-#[tokio::test]
-async fn fake_ledger_owns_document_status_and_cleanup_debt() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-    let status = DocumentStatus {
-        document_id: DocumentId::new("doc-a"),
-        source_id: SourceId::new("src_a"),
-        source_item_key: SourceItemKey::new("src/lib.rs"),
-        generation: SourceGenerationId::new("gen_1"),
-        status: DocumentLifecycleStatus::Published,
-        updated_at: ts(),
-        chunk_count: 1,
-        vector_point_count: 1,
-        error: None,
-        cleanup_status: None,
-    };
-
-    ledger.update_document_status(status.clone()).await.unwrap();
-    assert_eq!(
-        ledger.document_status(&DocumentId::new("doc-a")).await,
-        Some(status)
-    );
-    let updated = DocumentStatus {
-        document_id: DocumentId::new("doc-a"),
-        source_id: SourceId::new("src_a"),
-        source_item_key: SourceItemKey::new("src/lib.rs"),
-        generation: SourceGenerationId::new("gen_2"),
-        status: DocumentLifecycleStatus::Failed,
-        updated_at: ts_at(9),
-        chunk_count: 0,
-        vector_point_count: 0,
-        error: Some(SourceError {
-            code: "embed.failed".to_string(),
-            severity: Severity::Failed,
-            message: "embedding failed".to_string(),
-            source_item_key: Some(SourceItemKey::new("src/lib.rs")),
-            retryable: true,
-            provider_id: None,
-            cause: None,
-        }),
-        cleanup_status: Some(LifecycleStatus::Pending),
-    };
-    ledger
-        .update_document_status(updated.clone())
-        .await
-        .unwrap();
-    assert_eq!(
-        ledger.document_status(&DocumentId::new("doc-a")).await,
-        Some(updated)
-    );
-
-    ledger
-        .record_cleanup_debt(CleanupDebt {
-            debt_id: CleanupDebtId::new("debt-a"),
-            job_id: JobId::new(Uuid::from_u128(1)),
-            source_id: SourceId::new("src_a"),
-            generation: Some(SourceGenerationId::new("gen_1")),
-            kind: CleanupDebtKind::VectorDelete,
-            selector: CleanupSelector::Document {
-                document_id: DocumentId::new("doc-a"),
-            },
-            status: LifecycleStatus::Pending,
-            created_at: ts(),
-            attempts: 0,
-            last_error: None,
-            next_retry_at: None,
-            completed_at: None,
-        })
-        .await
-        .unwrap();
-    assert_eq!(ledger.cleanup_debt_count().await, 1);
-    ledger.reset().await.unwrap();
-    assert_eq!(ledger.cleanup_debt_count().await, 0);
-}
-
-#[tokio::test]
-async fn fake_rejects_document_status_and_cleanup_debt_for_missing_sources() {
-    let ledger = FakeLedgerStore::new();
-
-    let status_err = ledger
-        .update_document_status(DocumentStatus {
-            document_id: DocumentId::new("doc-missing"),
-            source_id: SourceId::new("missing"),
-            source_item_key: SourceItemKey::new("src/lib.rs"),
-            generation: SourceGenerationId::new("gen_1"),
-            status: DocumentLifecycleStatus::Published,
-            updated_at: ts(),
-            chunk_count: 1,
-            vector_point_count: 1,
-            error: None,
-            cleanup_status: None,
-        })
-        .await
-        .unwrap_err();
-    assert_eq!(status_err.code.to_string(), "source.ledger.source_missing");
-
-    let debt_err = ledger
-        .record_cleanup_debt(CleanupDebt {
-            debt_id: CleanupDebtId::new("debt-missing"),
-            job_id: JobId::new(Uuid::from_u128(1)),
-            source_id: SourceId::new("missing"),
-            generation: None,
-            kind: CleanupDebtKind::VectorDelete,
-            selector: CleanupSelector::Document {
-                document_id: DocumentId::new("doc-missing"),
-            },
-            status: LifecycleStatus::Pending,
-            created_at: ts(),
-            attempts: 0,
-            last_error: None,
-            next_retry_at: None,
-            completed_at: None,
-        })
-        .await
-        .unwrap_err();
-    assert_eq!(debt_err.code.to_string(), "source.ledger.source_missing");
-}
-
-#[tokio::test]
-async fn fake_rejects_cleanup_selector_source_or_generation_mismatch() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-
-    let base = CleanupDebt {
-        debt_id: CleanupDebtId::new("debt-mismatch"),
-        job_id: JobId::new(Uuid::from_u128(1)),
-        source_id: SourceId::new("src_a"),
-        generation: Some(SourceGenerationId::new("gen_1")),
-        kind: CleanupDebtKind::VectorDelete,
-        selector: CleanupSelector::SourceItem {
-            source_id: SourceId::new("other"),
-            source_item_key: SourceItemKey::new("src/lib.rs"),
-            generation: SourceGenerationId::new("gen_1"),
-        },
-        status: LifecycleStatus::Pending,
-        created_at: ts(),
-        attempts: 0,
-        last_error: None,
-        next_retry_at: None,
-        completed_at: None,
-    };
-    let source_err = ledger.record_cleanup_debt(base.clone()).await.unwrap_err();
-    assert_eq!(
-        source_err.code.to_string(),
-        "source.ledger.cleanup_selector_mismatch"
-    );
-
-    let mut generation_mismatch = base;
-    generation_mismatch.selector = CleanupSelector::SourceItem {
-        source_id: SourceId::new("src_a"),
-        source_item_key: SourceItemKey::new("src/lib.rs"),
-        generation: SourceGenerationId::new("gen_2"),
-    };
-    let generation_err = ledger
-        .record_cleanup_debt(generation_mismatch)
-        .await
-        .unwrap_err();
-    assert_eq!(
-        generation_err.code.to_string(),
-        "source.ledger.cleanup_selector_mismatch"
-    );
-}
-
-#[tokio::test]
-async fn fake_cleanup_debt_uses_natural_key_and_terminal_state_is_monotonic() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-
-    let mut debt = CleanupDebt {
-        debt_id: CleanupDebtId::new("debt-a"),
-        job_id: JobId::new(Uuid::from_u128(1)),
-        source_id: SourceId::new("src_a"),
-        generation: Some(SourceGenerationId::new("gen_1")),
-        kind: CleanupDebtKind::VectorDelete,
-        selector: CleanupSelector::Document {
-            document_id: DocumentId::new("doc-a"),
-        },
-        status: LifecycleStatus::Pending,
-        created_at: ts(),
-        attempts: 0,
-        last_error: None,
-        next_retry_at: None,
-        completed_at: None,
-    };
-
-    ledger.record_cleanup_debt(debt.clone()).await.unwrap();
-
-    debt.debt_id = CleanupDebtId::new("debt-b");
-    ledger.record_cleanup_debt(debt.clone()).await.unwrap();
-    assert_eq!(ledger.cleanup_debt_count().await, 1);
-
-    debt.status = LifecycleStatus::Completed;
-    debt.completed_at = Some(ts_at(10));
-    ledger.record_cleanup_debt(debt.clone()).await.unwrap();
-
-    debt.status = LifecycleStatus::Pending;
-    debt.completed_at = None;
-    ledger.record_cleanup_debt(debt).await.unwrap();
-
-    let stored = ledger
-        .cleanup_debt(&CleanupDebtId::new("debt-b"))
-        .await
-        .expect("stored debt");
-    assert_eq!(stored.status, LifecycleStatus::Completed);
-    assert_eq!(stored.completed_at, Some(ts_at(10)));
-}
-
-#[tokio::test]
-async fn fake_publish_creates_cleanup_debt_for_removed_items() {
-    let ledger = FakeLedgerStore::new();
-    ledger.upsert_source(source()).await.unwrap();
-
-    let gen1 = ledger
-        .create_generation(SourceId::new("src_a"))
-        .await
-        .unwrap();
-    ledger
-        .put_manifest(manifest_with_items(
-            &gen1.generation.0,
-            vec![
-                manifest_item("README.md", "same"),
-                manifest_item("src/old.rs", "removed"),
-            ],
-        ))
-        .await
-        .unwrap();
-    complete_and_publish(&ledger, completed_generation(gen1.clone())).await;
-
-    let gen2 = ledger
-        .create_generation(SourceId::new("src_a"))
-        .await
-        .unwrap();
-    ledger
-        .put_manifest(manifest_with_items(
-            &gen2.generation.0,
-            vec![manifest_item("README.md", "same")],
-        ))
-        .await
-        .unwrap();
-    let published = complete_and_publish(&ledger, completed_generation(gen2)).await;
-
-    assert_eq!(ledger.cleanup_debt_count().await, 1);
-    assert_eq!(published.publish_state, PublishState::CleanupPending);
-    assert_eq!(published.cleanup_debt.len(), 1);
-}
-
-#[tokio::test]
-async fn fake_ledger_acquires_conflicts_reclaims_and_releases_leases() {
-    let ledger = FakeLedgerStore::new();
-
-    let first = ledger
-        .acquire_lease(lease_request("source:src_a:refresh", "owner-a"))
-        .await
-        .unwrap()
-        .expect("first lease");
-    let conflict = ledger
-        .acquire_lease(lease_request("source:src_a:refresh", "owner-b"))
-        .await
-        .unwrap();
-    assert_eq!(conflict, None);
-
-    let wrong_owner_release = ledger
-        .release_lease(first.lease_id.clone(), "owner-b".to_string())
-        .await
-        .unwrap_err();
-    assert_eq!(
-        wrong_owner_release.code.to_string(),
-        "source.ledger.lease_owner_mismatch"
-    );
-
-    ledger
-        .release_lease(first.lease_id.clone(), "owner-a".to_string())
-        .await
-        .unwrap();
-    let expired = ledger
-        .acquire_lease(lease_request_ttl("source:src_a:refresh", "owner-a", 0))
-        .await
-        .unwrap()
-        .expect("zero ttl lease");
-    let reclaimed = ledger
-        .acquire_lease(lease_request("source:src_a:refresh", "owner-b"))
-        .await
-        .unwrap()
-        .expect("expired lease should be reclaimable");
-    assert_ne!(expired.lease_id, reclaimed.lease_id);
-    assert_eq!(reclaimed.owner_id, "owner-b");
-
-    let stale_heartbeat = ledger
-        .heartbeat_lease(expired.lease_id.clone(), "owner-a".to_string(), 30)
-        .await
-        .unwrap();
-    assert_eq!(stale_heartbeat, None);
-    let stale_release = ledger
-        .release_lease(expired.lease_id, "owner-a".to_string())
-        .await
-        .unwrap_err();
-    assert_eq!(
-        stale_release.code.to_string(),
-        "source.ledger.lease_missing"
-    );
-    let owner_c_conflict = ledger
-        .acquire_lease(lease_request("source:src_a:refresh", "owner-c"))
-        .await
-        .unwrap();
-    assert_eq!(owner_c_conflict, None);
-
-    ledger
-        .release_lease(reclaimed.lease_id.clone(), "owner-b".to_string())
-        .await
-        .unwrap();
-    let reacquired = ledger
-        .acquire_lease(lease_request("source:src_a:refresh", "owner-a"))
-        .await
-        .unwrap();
-    assert!(reacquired.is_some());
-}
-
-#[tokio::test]
-async fn fake_ledger_same_owner_can_renew_lease() {
-    let ledger = FakeLedgerStore::new();
-    let first = ledger
-        .acquire_lease(lease_request("source:src_a:refresh", "owner-a"))
-        .await
-        .unwrap()
-        .expect("first lease");
-
-    let renewed = ledger
-        .acquire_lease(lease_request("source:src_a:refresh", "owner-a"))
-        .await
-        .unwrap()
-        .expect("same owner renewal");
-
-    assert_eq!(renewed.lease_id, first.lease_id);
-    assert_eq!(renewed.acquired_at, first.acquired_at);
-}
-
-#[tokio::test]
-async fn fake_ledger_heartbeat_rejects_expired_lease() {
-    let ledger = FakeLedgerStore::new();
-    let first = ledger
-        .acquire_lease(lease_request_ttl("source:src_a:refresh", "owner-a", 0))
-        .await
-        .unwrap()
-        .expect("first lease");
-
-    let heartbeat = ledger
-        .heartbeat_lease(first.lease_id, "owner-a".to_string(), 30)
-        .await
-        .unwrap();
-
-    assert_eq!(heartbeat, None);
-}
-
-#[tokio::test]
-async fn fake_ledger_heartbeat_rejects_wrong_owner() {
-    let ledger = FakeLedgerStore::new();
-    let first = ledger
-        .acquire_lease(lease_request("source:src_a:refresh", "owner-a"))
-        .await
-        .unwrap()
-        .expect("first lease");
-
-    let heartbeat = ledger
-        .heartbeat_lease(first.lease_id, "owner-b".to_string(), 30)
-        .await
-        .unwrap();
-
-    assert_eq!(heartbeat, None);
-}
+#[path = "store_tests/store_document_cleanup_tests.rs"]
+mod store_document_cleanup_tests;
+#[path = "store_tests/store_lease_tests.rs"]
+mod store_lease_tests;
+#[path = "store_tests/store_manifest_tests.rs"]
+mod store_manifest_tests;
